+++ CHANGELOG.rst
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is inspired by `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-[0.3.0] - 2025-01-07
+[0.3.0] - 2025-01-08
 --------------------
+Added
+^^^^^
+Updated APILoader to load the checklists for a sub-region if the number of results
+returned matches the API's result limit.
+
 Changed
 ^^^^^^^
 Refactored APILoader to make it easier to reuse.